move imported module to recipe
